+++ aws/init.go
@@ -32,7 +32,7 @@ import (
 )
 
 var (
-	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService, LambdaService, MonitoringService cloud.Service
+	AccessService, InfraService, StorageService, NotificationService, QueueService, DnsService, LambdaService, MonitoringService, CdnService cloud.Service
 )
 
 func InitServices(conf map[string]interface{}, log *logger.Logger) error {
@@ -55,6 +55,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	DnsService = NewDns(sess, awsconf, log)
 	LambdaService = NewLambda(sess, awsconf, log)
 	MonitoringService = NewMonitoring(sess, awsconf, log)
+	CdnService = NewCdn(sess, awsconf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -64,6 +65,7 @@ func InitServices(conf map[string]interface{}, log *logger.Logger) error {
 	cloud.ServiceRegistry[DnsService.Name()] = DnsService
 	cloud.ServiceRegistry[LambdaService.Name()] = LambdaService
 	cloud.ServiceRegistry[MonitoringService.Name()] = MonitoringService
+	cloud.ServiceRegistry[CdnService.Name()] = CdnService
 
 	return nil
 }
@@ -96,6 +98,7 @@ func NewDriver(region, profile string, log ...*logger.Logger) (driver.Driver, er
 	drivers = append(drivers, NewDns(sess, awsconf, drivLog).Drivers()...)
 	drivers = append(drivers, NewLambda(sess, awsconf, drivLog).Drivers()...)
 	drivers = append(drivers, NewMonitoring(sess, awsconf, drivLog).Drivers()...)
+	drivers = append(drivers, NewCdn(sess, awsconf, drivLog).Drivers()...)
 
 	return driver.NewMultiDriver(drivers...), nil
 }