+++ graph/rdf.go
@@ -69,6 +69,17 @@ func getPropertyValue(gph tstore.RDFGraph, propObj tstore.Object, prop string) (
 			return nil, err
 		}
 		return kv, nil
+	case definedBy == rdf.RdfsList && dataType == rdf.DistributionOrigin:
+		id, ok := propObj.Resource()
+		if !ok {
+			return nil, fmt.Errorf("get property '%s': object not resource identifier", prop)
+		}
+		o := &DistributionOrigin{}
+		err := o.unmarshalFromTriples(gph, id)
+		if err != nil {
+			return nil, err
+		}
+		return o, nil
 	default:
 		return "", fmt.Errorf("get property value: %s is neither literal nor class, nor list", definedBy)
 	}